refactor dftx_pool.test.ts to PoolSwap.test.ts
